backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
django==5.2.6
django-filter==25.2
djangorestframework==3.16.1
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.28.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.2
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
